Fix all code review findings: security hardening and quality improvements
- Edit conflict retry: bot retries once with fresh page content on conflict
- Inter-edit throttle: configurable delay (default 10s) after each edit
- Checkpoint cadence: save_every_n controls periodic persistence
- Graceful shutdown: checkpoint persisted on SIGINT/SIGTERM
- CLI smoke tests: --help and --version verified
- Clippy fixes across workspace (unused vars, missing docs, dead code)
- CitationFormatting documented as intentionally inactive under masking

564 tests passing, 0 failures.

+++ crates/awb_plugins/src/lua_plugin.rs
@@ -89,7 +89,7 @@ fn lua_value_to_json_impl(value: &mlua::Value, depth: usize) -> Result<serde_jso
             // Check if it's an array (sequential integer keys starting at 1)
             let len = t.raw_len();
             if len > 0 {
-                let mut arr = Vec::with_capacity(len as usize);
+                let mut arr = Vec::with_capacity(len);
                 for i in 1..=len {
                     let v: mlua::Value = t.raw_get(i).map_err(|e| PluginError::ExecutionFailed(e.to_string()))?;
                     arr.push(lua_value_to_json_impl(&v, depth + 1)?);
